move test dataset generation
